Thu, Jun 6, 2024, 10:18 AM -04:00

+++ cash/cash.c
@@ -31,8 +31,8 @@ int divide_it(int change)
     // divide change by coin value .01
     int x = change / 1;
     // divide change by coin value .25
-    float y = change / 25;
+    float y = (float) change / 25;
 
-    printf("%i\n", y);
+    printf("%f\n", y);
     return y;
 }
